# api/openapi.yaml
openapi: 3.1.0
info:
  title: HeatCheck Public API
  version: 1.0.0
  description: |
    Production API for HeatCheck challenges, proof clips, OpenAI-assisted
    verification, paid participation, voting, moderation, account lifecycle,
    and copyright workflows. All timestamps use RFC 3339 UTC values.
servers:
  - url: https://heatcheck.dogi.watch
    description: Production
  - url: http://localhost:8080
    description: Local development
...
security: []
paths:
...
  /v1/me:
    get:
      tags: [Authentication]
      operationId: getMe
      summary: Get the authenticated account
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current account and missing policy acceptances.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: object
                    required: [user, missing_policy_acceptances]
                    properties:
                      user:
                        $ref: "#/components/schemas/PrivateUser"
                      missing_policy_acceptances:
                        type: array
                        items:
                          $ref: "#/components/schemas/PolicyAcceptance"
        "401":
          $ref: "#/components/responses/Error"
    delete:
      tags: [Accounts]
      operationId: requestAccountDeletion
      summary: Schedule account deletion
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties:
                password:
                  type: string
                  writeOnly: true
      responses:
        "202":
          description: Account deletion scheduled after the configured grace period.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeletionEnvelope"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
...
  /v1/me/exports:
    post:
      tags: [Accounts]
      operationId: createAccountExport
      summary: Request an account export
      security:
        - bearerAuth: []
      responses:
        "202":
          description: Export queued, or existing active export returned.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountExportEnvelope"
        "403":
          $ref: "#/components/responses/Error"
        "429":
          $ref: "#/components/responses/Error"
  /v1/me/exports/{exportID}:
    parameters:
      - $ref: "#/components/parameters/ExportID"
    get:
      tags: [Accounts]
      operationId: getAccountExport
      summary: Get account export status
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Export status and a short-lived download URL when ready.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountExportEnvelope"
        "404":
          $ref: "#/components/responses/Error"
  /v1/me/deletion:
    get:
      tags: [Accounts]
      operationId: getAccountDeletion
      summary: Get pending deletion status
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Pending deletion schedule.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeletionEnvelope"
        "404":
          $ref: "#/components/responses/Error"
    delete:
      tags: [Accounts]
      operationId: cancelAccountDeletion
      summary: Cancel account deletion
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Pending account deletion cancelled.
        "404":
          $ref: "#/components/responses/Error"
...
  /v1/copyright/notices:
    post:
      tags: [Copyright]
      operationId: createCopyrightNotice
      summary: Submit a copyright notice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCopyrightNotice"
      responses:
        "201":
          description: Notice received. Sensitive claimant data is not echoed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CopyrightReceiptEnvelope"
        "422":
          $ref: "#/components/responses/Error"
        "429":
          $ref: "#/components/responses/Error"
...
  /v1/admin/copyright/notices:
    get:
      tags: [Administration]
      operationId: adminListCopyrightNotices
      summary: List copyright notices
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [received, reviewing, actioned, rejected, countered, restored, closed]
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Copyright notice queue including private claimant data.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CopyrightNoticeListEnvelope"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Offset:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
...
    ExportID:
      name: exportID
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    Error:
      description: Structured API error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details: {}
    Pagination:
      type: object
      required: [limit, offset]
      properties:
        limit:
          type: integer
        offset:
          type: integer
    PolicyAcceptance:
      type: object
      required: [kind, version]
      properties:
        kind:
          type: string
        version:
          type: string
...
    PublicUser:
      type: object
      required: [id, handle, display_name, created_at]
      properties:
        id:
          type: string
          format: uuid
        handle:
          type: string
        display_name:
          type: string
        created_at:
          type: string
          format: date-time
    PrivateUser:
      allOf:
        - $ref: "#/components/schemas/PublicUser"
        - type: object
          required: [email, date_of_birth, role, status]
          properties:
            email:
              type: string
              format: email
            date_of_birth:
              type: string
              format: date
            role:
              type: string
              enum: [user, moderator, admin]
            status:
              type: string
              enum: [active, suspended, deletion_pending, deleted]
            email_verified_at:
              type: [string, "null"]
              format: date-time
...
    AccountExport:
      type: object
      required: [id, status, created_at]
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, processing, ready, failed, expired]
        download_url:
          type: string
          format: uri
        expires_at:
          type: [string, "null"]
          format: date-time
        error:
          type: string
        created_at:
          type: string
          format: date-time
        completed_at:
          type: [string, "null"]
          format: date-time
    AccountExportEnvelope:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/AccountExport"
...
    DeletionEnvelope:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [status, execute_after]
          properties:
            status:
              type: string
              const: deletion_pending
            execute_after:
              type: string
              format: date-time
...
    CreateCopyrightNotice:
      type: object
      required:
        - claimant_name
        - claimant_email
        - claimant_address
        - relationship
        - copyrighted_work
        - infringing_url
        - good_faith
        - accuracy
        - signature
      properties:
        claimant_name:
          type: string
        claimant_email:
          type: string
          format: email
        claimant_address:
          type: string
        relationship:
          type: string
        copyrighted_work:
          type: string
        infringing_url:
          type: string
          format: uri
        submission_id:
          type: [string, "null"]
          format: uuid
        good_faith:
          type: boolean
          const: true
        accuracy:
          type: boolean
          const: true
        signature:
          type: string
...
    CopyrightReceiptEnvelope:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [id, status, created_at]
          properties:
            id:
              type: string
              format: uuid
            notice_id:
              type: string
              format: uuid
            status:
              type: string
            created_at:
              type: string
              format: date-time
    CopyrightNotice:
      type: object
      required:
        - id
        - claimant_name
        - claimant_email
        - claimant_address
        - relationship
        - copyrighted_work
        - infringing_url
        - good_faith
        - accuracy
        - signature
        - status
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        claimant_name:
          type: string
        claimant_email:
          type: string
          format: email
        claimant_address:
          type: string
        relationship:
          type: string
        copyrighted_work:
          type: string
        infringing_url:
          type: string
          format: uri
        submission_id:
          type: [string, "null"]
          format: uuid
        good_faith:
          type: boolean
        accuracy:
          type: boolean
        signature:
          type: string
        status:
          type: string
          enum: [received, reviewing, actioned, rejected, countered, restored, closed]
        resolution_note:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        actioned_at:
          type: [string, "null"]
          format: date-time
        counter_notice_due:
          type: [string, "null"]
          format: date-time
...
    CopyrightNoticeListEnvelope:
      type: object
      required: [data, pagination]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/CopyrightNotice"
        pagination:
          $ref: "#/components/schemas/Pagination"
...